+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 <!-- TODO: add previous versions -->
+## v1.0.3
+
+- Update rocker/shiny to v4.5.1
+- Add waiter
+- Add links to external resources
+- Fix data downloads
+
 ## v1.0.2
 
 - Bug fixes in table export